STM32F0: Fix the PLL multiplier table and fix the 48MHz PLL setting
The value '6' was twice in the table and all higher frequencies are
shifted. The values are now fitting the table in 'STM32F05xxx/06xxx advanced
ARM-based 32-bit MCUs', page 101.

Also, the 48MHz PLL setting used the wrong multiplier and was actually
set to overclocking the device. The multiplier is set to x 6 now, for
proper 48MHz.

The PLL frequencies have been measured by selecting

    rcc_set_mco(RCC_CFGR_MCO_SYSCLK);

and measuring the output with an oscilloscope. 8, 16, 24, 32, 40 and 48 MHz
work fine from the HSI base.